Moved IRegisteredNodesResponse and associated definitions to HederaSchema.ts
