+++ front-end/src/shared/interfaces/index.ts
@@ -1,6 +1,5 @@
 export * from './IAccountInfoParsed';
 export * from './INodeInfoParsed';
-export * from './IRegisteredNode';
 export * from './HederaSchema';
 export * from './Theme';
 export * from './organization';
